stream_sidebar: Implement stream navigation behavior.
Clicking on the name of a stream in the left sidebar
now navigates to the top topic in the left sidebar
view of that stream, rather than an interleaved view.
Added an "interleaved" button to the stream header
row in the left sidebar that appears only when the
user hovers over it.

Additionally, update the navigation.tests.ts and
message-basics.test.ts to incorporate this
feature.

Fixes #26937.